初版保存

+++ src/main/java/com/songmin/service/imp/FileOperateServiceImp.java
@@ -4,6 +4,7 @@ import com.songmin.dao.UserOperateMapper;
 import com.songmin.model.ResultMap;
 import com.songmin.model.User;
 import com.songmin.service.FileOperateService;
+import com.songmin.utils.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,20 +27,13 @@ public class FileOperateServiceImp implements FileOperateService {
     private static final String basePath = "D:\\Projects\\Work\\流浪动物救助\\Files\\"; //文件存储基础路径
 
     @Override
-    public ResultMap<Map<String, String>> uploadAvatar(MultipartFile file, String token) {
+    public ResultMap<Map<String, String>> uploadAvatar(MultipartFile file, String userId) {
         ResultMap<Map<String, String>> result = new ResultMap<>();
         if (file.isEmpty()) {
             result.setCode(400);
             return result;
         }
-        String userId = redisTemplate.opsForValue().get(token).toString();
-        List<User> users = userOperateMapper.queryUserInfoById(userId);
-        if (users == null || users.size() < 1) {
-            result.setCode(404);
-            return result;
-        }
-        User user = users.get(0);
-        String fileBasePath = basePath + user.getCount() + "\\" + user.getFilePath() + "\\" + "images\\";
+        String fileBasePath = basePath + userId + "\\";
         try {
             File path = new File(fileBasePath);
             if (!path.exists()) {
@@ -47,13 +41,15 @@ public class FileOperateServiceImp implements FileOperateService {
             }
             // 删除原有文件
             for (File f : path.listFiles()) {
-                f.delete();
+                if (f.getName().contains("avatar")) {
+                    f.delete();
+                }
             }
-            File avatarFile = new File(fileBasePath + file.getOriginalFilename());
+            String avatarName = file.getOriginalFilename();
+            avatarName = avatarName.substring(avatarName.lastIndexOf("."));
+            avatarName = "avatar" + avatarName;
+            File avatarFile = new File(fileBasePath + avatarName);
             file.transferTo(avatarFile);
-            user.setAvatar(file.getOriginalFilename());
-            user.setUserId(userId);
-            userOperateMapper.updateUserInfo(user);
             logger.info(file.getOriginalFilename() + "上传成功!");
         } catch (Exception e) {
             e.printStackTrace();
@@ -63,43 +59,25 @@ public class FileOperateServiceImp implements FileOperateService {
     }
 
     @Override
-    public void downloadAvatar(String token, HttpServletResponse response) throws Exception {
-        token = token.replaceAll("Bearer", "");
-        String userId = redisTemplate.opsForValue().get(token).toString();
-        List<User> users = userOperateMapper.queryUserInfoById(userId);
-        if (users == null || users.size() < 1) {
-            response.sendError(404);
-            return;
-        }
-        User user = users.get(0);
-        String avatar = user.getAvatar();
-        String contentType = "image/";
-        String fileBasePath = basePath;
-        if (avatar == null || "".equals(avatar)) {
-            fileBasePath += "public\\images\\photo_default.jpg";
-            contentType += "jpg";
-        } else {
-            fileBasePath += user.getCount() + "\\" + user.getFilePath() + "\\" + "images\\" + user.getAvatar();
-            contentType += avatar.substring(avatar.lastIndexOf(".") + 1);
-        }
-        File avatarPath = new File(fileBasePath);
-        if (!avatarPath.exists()) {
-            response.sendError(404);
-            return;
-        }
-        response.setContentType(contentType);
-        //response.setCharacterEncoding("UTF-8");
-        response.setHeader("Content-Disposition", "attachment;fileName=" + avatar);
-        byte[] buffer = new byte[1024];
-        try (BufferedInputStream bis = new BufferedInputStream(new FileInputStream(avatarPath))) {
-            OutputStream os = response.getOutputStream();
-            int i = bis.read(buffer);
-            while (i != -1) {
-                os.write(buffer);
-                i = bis.read(buffer);
+    public void downloadAvatar(String userId, HttpServletResponse response) {
+        String avatarName = "";
+        String filePath = basePath + userId + "\\";
+        File file = new File(filePath);
+        for (File f : file.listFiles()) {
+            if (f.getName().contains("avatar")){
+                avatarName = f.getName();
+                break;
             }
-            os.flush();
-        } catch (IOException e) {
+        }
+        if (avatarName == null || "".equals(avatarName)) {
+            filePath += "public\\images\\photo_default.jpg";
+        } else {
+            filePath += avatarName;
+        }
+
+        try {
+            FileUtils.sendFileStream(filePath, response);
+        } catch (Exception e) {
             e.printStackTrace();
         }
     }
@@ -108,6 +86,8 @@ public class FileOperateServiceImp implements FileOperateService {
     public ResultMap<Map<String, String>> uploadRescuePhoto(MultipartFile file, String userId) {
         ResultMap<Map<String, String>> resultMap = new ResultMap();
         Map<String, String> message = new HashMap<>();
+        //将上传文件缓存，等待用户确认提交
+        List<String> cachedUploadFileList;
 
         if (file.isEmpty()) {
             message.put("msg", "上传文件为空!");
@@ -115,16 +95,17 @@ public class FileOperateServiceImp implements FileOperateService {
             resultMap.setResult(message);
             return resultMap;
         }
-        List<User> userList = userOperateMapper.queryUserInfoById(userId);
-        if (userList == null || userList.size() == 0) {
-            message.put("msg", "获取用户信息失败!");
-            resultMap.setCode(400);
-            resultMap.setResult(message);
-            return resultMap;
+
+        //获取已缓存的待上传文件列表
+        Object obj = redisTemplate.opsForHash().get("rescue_upload", userId);
+        if (obj != null) {
+            cachedUploadFileList = (ArrayList<String>)obj;
+        } else {
+            cachedUploadFileList = new ArrayList<>();
         }
-        User user = userList.get(0);
+
         try {
-            String baseFilePath = basePath + user.getCount() + "\\" + user.getFilePath() + "\\apply\\";
+            String baseFilePath = basePath + userId + "\\apply\\rescue\\";
             File basePathFile = new File(baseFilePath);
             if (!basePathFile.exists()) {
                 basePathFile.mkdirs();
@@ -132,6 +113,9 @@ public class FileOperateServiceImp implements FileOperateService {
             String filePath = baseFilePath + file.getOriginalFilename();
             File saveFile = new File(filePath);
             file.transferTo(saveFile);
+            //保存新增文件信息
+            cachedUploadFileList.add(filePath);
+            redisTemplate.opsForHash().put("rescue_upload", userId, cachedUploadFileList);
         } catch (IOException e) {
             e.printStackTrace();
         }
@@ -146,33 +130,97 @@ public class FileOperateServiceImp implements FileOperateService {
     public ResultMap<Map<String, String>> deleteRescuePhoto(String userId, String fileName) {
         ResultMap<Map<String, String>> resultMap = new ResultMap();
         Map<String, String> message = new HashMap<>();
-        List<User> userList = userOperateMapper.queryUserInfoById(userId);
+        //将删除文件信息缓存，等待用户确认
+        List<String> cachedDeleteFileList;
 
-        if (userList == null || userList.size() == 0) {
-            message.put("msg", "获取用户信息失败!");
-            resultMap.setCode(400);
-            resultMap.setResult(message);
-            return resultMap;
+        //获取已缓存待删除文件信息
+        Object obj = redisTemplate.opsForHash().get("rescue_delete", userId);
+        if (obj != null) {
+            cachedDeleteFileList = (ArrayList<String>)obj;
+        } else {
+            cachedDeleteFileList = new ArrayList<>();
         }
-        User user = userList.get(0);
-        String baseFilePath = basePath + user.getCount() + "\\" + user.getFilePath() + "\\apply\\";
-        File basePathFile = new File(baseFilePath);
-        if (!basePathFile.exists()) {
-            message.put("msg", "图片路径不存在!");
-            resultMap.setCode(400);
-            resultMap.setResult(message);
-            return resultMap;
+
+        String baseFilePath = basePath + userId + "\\apply\\rescue\\" + fileName;
+        //保存新增待删除文件信息
+        cachedDeleteFileList.add(baseFilePath);
+        redisTemplate.opsForHash().put("rescue_delete", userId, cachedDeleteFileList);
+        message.put("msg", "OK");
+        resultMap.setCode(200);
+        resultMap.setResult(message);
+
+        return resultMap;
+    }
+
+    @Override
+    public ResultMap<Map<String, String>> submitRescuePhoto(String userId) {
+        ResultMap<Map<String, String>> resultMap = new ResultMap<>();
+        Map<String, String> message = new HashMap<>();
+        List<String> cachedFileList;
+
+        //获取缓存的上传图片
+        Object obj = redisTemplate.opsForHash().get("rescue_upload", userId);
+        if (obj != null) {
+            //清除上传缓存信息
+            redisTemplate.opsForHash().delete("rescue_upload", userId);
         }
-        for (File f : basePathFile.listFiles()) {
-            if (f.getName().contains(fileName)) {
-                f.delete();
+        //获取缓存的删除图片信息
+        obj = redisTemplate.opsForHash().get("rescue_delete", userId);
+        if (obj != null) {
+            cachedFileList = (ArrayList<String>)obj;
+            File file;
+            for (String fileName : cachedFileList) {
+                file = new File(fileName);
+                if (file.exists()) {
+                    file.delete();
+                }
             }
+            //清除缓存信息
+            redisTemplate.opsForHash().delete("rescue_delete", userId);
         }
         message.put("msg", "OK");
         resultMap.setCode(200);
         resultMap.setResult(message);
 
         return resultMap;
+    }
+
+    @Override
+    public ResultMap<Map<String, String>> discardRescuePhoto(String userId) {
+        ResultMap<Map<String, String>> resultMap = new ResultMap<>();
+        Map<String, String> message = new HashMap<>();
+        List<String> cachedFileList;
+
+        Object obj = redisTemplate.opsForHash().get("rescue_upload", userId);
+        if (obj != null) {
+            cachedFileList = (ArrayList<String>)obj;
+            File file;
+            for (String fileName : cachedFileList) {
+                file = new File(fileName);
+                if (file.exists()) {
+                    file.delete();
+                }
+            }
+            redisTemplate.opsForHash().delete("rescue_upload", userId);
+        }
+
+        obj = redisTemplate.opsForHash().get("rescue_delete", userId);
+        if (obj != null) {
+            redisTemplate.opsForHash().delete("rescue_delete", userId);
+        }
+
+        return resultMap;
+    }
+
+    @Override
+    public void downloadRescuePhoto(String userId, String fileName, HttpServletResponse response) {
+        String filePath = basePath + userId + "\\apply\\rescue\\" + fileName;
+
+        try {
+            FileUtils.sendFileStream(filePath, response);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
     }
 
 }
